# telemetry: perf: Use `metric.WithAttributeSet`