--- a/pubsublite/internal/telemetry/publisher.go
+++ b/pubsublite/internal/telemetry/publisher.go
@@ -117,14 +117,14 @@ func (p *Producer) Publish(ctx context.Context, msg *pubsub.Message) pubsubabs.P
 		if err != nil {
 			span.RecordError(err)
 			span.SetStatus(codes.Error, err.Error())
-			p.metrics.errored.Add(context.Background(), 1, metric.WithAttributes(
-				attrs...,
-			))
+			p.metrics.errored.Add(context.Background(), 1,
+				metric.WithAttributeSet(attribute.NewSet(attrs...)),
+			)
 		} else {
 			span.SetStatus(codes.Ok, "success")
-			p.metrics.produced.Add(context.Background(), 1, metric.WithAttributes(
-				attrs...,
-			))
+			p.metrics.produced.Add(context.Background(), 1,
+				metric.WithAttributeSet(attribute.NewSet(attrs...)),
+			)
 		}
 		span.End()
 	}()